Split controller.go off of the KPA reconciler
